pyproject.toml
[project]
name = "industryproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "mysql-connector-python>=9.6.0",
]
